Split notification functions into multiple files
