+++ packages/rocketchat-lib/server/lib/roomTypes.js
@@ -31,6 +31,9 @@ RocketChat.roomTypes = new class roomTypesServer extends RoomTypesCommon {
 		return this.roomTypes[roomType] && this.roomTypes[roomType].roomFind;
 	}
 
+	getRoomName(roomType, roomData) {
+		return this.roomTypes[roomType] && this.roomTypes[roomType].roomName && this.roomTypes[roomType].roomName(roomData);
+	}
 
 	/**
 	 * Run the publish for a room type
